server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0
httpx==0.27.0
twilio==9.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
edge-tts==6.1.12
